refactor: moved chat related models into a separate folder
