+++ Sources/OpenAI/Public/Models/Chat/ChatFunction.swift
@@ -1,0 +1,38 @@
+//
+//  ChatFunction.swift
+//  
+//
+//  Created by Federico Vitale on 14/11/23.
+//
+
+import Foundation
+
+public struct ChatFunctionCall: Codable, Equatable {
+    /// The name of the function to call.
+    public let name: String?
+    /// The arguments to call the function with, as generated by the model in JSON format. Note that the model does not always generate valid JSON, and may hallucinate parameters not defined by your function schema. Validate the arguments in your code before calling your function.
+    public let arguments: String?
+
+    public init(name: String?, arguments: String?) {
+        self.name = name
+        self.arguments = arguments
+    }
+}
+
+public struct ChatFunctionDeclaration: Codable, Equatable {
+    /// The name of the function to be called. Must be a-z, A-Z, 0-9, or contain underscores and dashes, with a maximum length of 64.
+    public let name: String
+    
+    /// The description of what the function does.
+    public let description: String
+    
+    /// The parameters the functions accepts, described as a JSON Schema object.
+    public let parameters: JSONSchema
+  
+    public init(name: String, description: String, parameters: JSONSchema) {
+      self.name = name
+      self.description = description
+      self.parameters = parameters
+    }
+}
+
