+++ Sources/OpenAI/Public/Models/Chat/ChatResult.swift
