+++ Sources/OpenAI/Public/Models/Chat/ChatStreamResult.swift
